pyproject.toml
[tool.poetry]
name = "bottless"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
zappa = "^0.53.0"
bottle = "^0.12.19"
boto3 = "^1.18.26"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
